Extract waitForReadiness function
