+++ src/wait-for-readiness/index.js
@@ -1,0 +1,1 @@
+export { default } from './wait-for-readiness'
